requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.3.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Migrate==4.0.4
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dotenv==1.0.0
pytz==2024.2
referencing==0.35.1
rpds-py==0.22.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
